+++ backend/btrixcloud/storages.py
@@ -5,7 +5,18 @@ from typing import Union
 from urllib.parse import urlsplit
 from contextlib import asynccontextmanager
 
+import asyncio
+import json
+import queue
+import time
+
+from datetime import datetime
+
 from fastapi import Depends, HTTPException
+from stream_zip import stream_zip, NO_COMPRESSION_64
+
+from boto3.s3.transfer import TransferConfig
+
 
 import aiobotocore.session
 import botocore.session
@@ -14,6 +25,9 @@ import boto3
 from .orgs import Organization, DefaultStorage, S3Storage
 from .users import User
 from .zip import get_zip_file, extract_and_parse_log_file
+
+
+CHUNK_SIZE = 1024 * 256
 
 
 # ============================================================================
@@ -305,10 +319,18 @@ async def get_presigned_url(org, crawlfile, crawl_manager, duration=3600):
 # ============================================================================
 async def delete_crawl_file_object(org, crawlfile, crawl_manager):
     """delete crawl file from storage."""
+    return await delete_file(
+        org, crawlfile.filename, crawl_manager, crawlfile.def_storage_name
+    )
+
+
+# ============================================================================
+async def delete_file(org, filename, crawl_manager, def_storage_name="default"):
+    """delete specified file from storage"""
     status_code = None
 
-    if crawlfile.def_storage_name:
-        s3storage = await crawl_manager.get_default_storage(crawlfile.def_storage_name)
+    if def_storage_name:
+        s3storage = await crawl_manager.get_default_storage(def_storage_name)
 
     elif org.storage.type == "s3":
         s3storage = org.storage
@@ -321,11 +343,11 @@ async def delete_crawl_file_object(org, crawlfile, crawl_manager):
         bucket,
         key,
     ):
-        key += crawlfile.filename
+        key += filename
         response = await client.delete_object(Bucket=bucket, Key=key)
         status_code = response["ResponseMetadata"]["HTTPStatusCode"]
 
-    return status_code
+    return status_code == 204
 
 
 # ============================================================================
@@ -364,6 +386,7 @@ async def get_wacz_logs(org, crawlfile, crawl_manager):
         return combined_log_lines
 
 
+# ============================================================================
 def get_public_policy(bucket_path):
     """return public policy for /public paths"""
     return {
@@ -378,3 +401,223 @@ def get_public_policy(bucket_path):
             }
         ],
     }
+
+
+# ============================================================================
+def _sync_dl(all_files, client, bucket, key):
+    """generate streaming zip as sync"""
+    for file_ in all_files:
+        file_.path = file_.name
+
+    datapackage = {
+        "profile": "multi-wacz-package",
+        "resources": [file_.dict() for file_ in all_files],
+    }
+    datapackage = json.dumps(datapackage).encode("utf-8")
+
+    def get_file(name):
+        response = client.get_object(Bucket=bucket, Key=key + name)
+        return response["Body"].iter_chunks(chunk_size=CHUNK_SIZE)
+
+    def member_files():
+        modified_at = datetime(year=1980, month=1, day=1)
+        perms = 0o664
+        for file_ in all_files:
+            yield (
+                file_.name,
+                modified_at,
+                perms,
+                NO_COMPRESSION_64,
+                get_file(file_.name),
+            )
+
+        yield (
+            "datapackage.json",
+            modified_at,
+            perms,
+            NO_COMPRESSION_64,
+            (datapackage,),
+        )
+
+    return stream_zip(member_files(), chunk_size=CHUNK_SIZE)
+
+
+# ============================================================================
+async def download_streaming_wacz(org, crawl_manager, files):
+    """return an iter for downloading a stream nested wacz file
+    from list of files"""
+    client, bucket, key, _ = await get_sync_client(org, crawl_manager)
+
+    loop = asyncio.get_event_loop()
+
+    resp = await loop.run_in_executor(None, _sync_dl, files, client, bucket, key)
+
+    return resp
+
+
+# ============================================================================
+async def upload_streaming_wacz(
+    target_path, org, crawl_manager, files, update_published_callback
+):
+    """perform a streaming upload of a wacz file for the specified collection"""
+
+    client, bucket, key, endpoint_url = await get_sync_client(
+        org, crawl_manager, use_full=True
+    )
+
+    # set published url to empty / publishing
+    await update_published_callback("")
+
+    loop = asyncio.get_event_loop()
+
+    msgq = queue.Queue()
+
+    total_size = 0
+    for file_ in files:
+        total_size += file_.size
+
+    uploading = True
+
+    async def process_q(msgq, total_size, update_published_callback):
+        """update upload size in db"""
+
+        while uploading:
+            try:
+                new_value = msgq.get_nowait()
+                if new_value == -1:
+                    break
+
+                percent = 100 * new_value / total_size
+
+                await update_published_callback("", percent)
+
+            except queue.Empty:
+                await asyncio.sleep(3)
+
+    async def finish_task(published_url):
+        if not await loop.run_in_executor(
+            None,
+            _sync_publish,
+            files,
+            client,
+            bucket,
+            key,
+            target_path,
+            msgq,
+        ):
+            # publishing failed
+            published_url = ""
+
+        nonlocal uploading
+        uploading = False
+
+        await update_published_callback(published_url)
+
+    asyncio.create_task(process_q(msgq, total_size, update_published_callback))
+
+    published_url = endpoint_url + target_path
+    asyncio.create_task(finish_task(published_url))
+
+    return {"publishing": True}
+
+
+# ============================================================================
+def _sync_publish(all_files, client, bucket, key, target_path, msgq):
+    """publish collection to public s3 path"""
+
+    counter = UploadCounter(msgq)
+    try:
+        target_path = key + target_path
+
+        wacz_stream = _sync_dl(all_files, client, bucket, key)
+        wacz_stream = to_file_like_obj(wacz_stream)
+
+        # set part size to 5MB
+        config = TransferConfig(multipart_threshold=5 * 1024**2)
+
+        client.upload_fileobj(
+            Fileobj=wacz_stream,
+            Bucket=bucket,
+            Key=target_path,
+            Callback=counter.update,
+            Config=config,
+        )
+
+        bucket_path = bucket + "/" + key.rstrip("/") if key else bucket
+
+        policy = json.dumps(get_public_policy(bucket_path))
+
+        client.put_bucket_policy(Bucket=bucket, Policy=policy)
+
+        # indicate we're done with this q
+        msgq.put(-1)
+
+        return True
+
+    # pylint: disable=broad-exception-caught
+    except Exception:
+        return False
+
+
+# ============================================================================
+# pylint: disable=too-few-public-methods
+class UploadCounter:
+    """UploadCounter"""
+
+    def __init__(self, msgq):
+        self.counter = 0
+        self.msgq = msgq
+        self.last_update = 0
+
+    def update(self, num):
+        """upload callback"""
+        self.counter += num
+        update_time = time.time()
+        if (update_time - self.last_update) > 1:
+            self.last_update = update_time
+            self.msgq.put(self.counter)
+
+
+# ============================================================================
+def to_file_like_obj(iterable):
+    """iter to file like obj"""
+    chunk = b""
+    offset = 0
+    # pylint: disable=invalid-name
+    it = iter(iterable)
+
+    def up_to_iter(size):
+        nonlocal chunk, offset
+
+        # if no size, yield exactly one chunk
+        if not size or size < 0:
+            try:
+                chunk = next(it)
+                yield chunk
+            except StopIteration:
+                pass
+
+            return
+
+        while size:
+            if offset == len(chunk):
+                try:
+                    chunk = next(it)
+                except StopIteration:
+                    break
+                else:
+                    offset = 0
+            to_yield = min(size, len(chunk) - offset)
+            offset = offset + to_yield
+            size -= to_yield
+            yield chunk[offset - to_yield : offset]
+
+    # pylint: disable=too-few-public-methods
+    class FileLikeObj:
+        """file-like obj wrapper for upload"""
+
+        def read(self, size=-1):
+            """read interface for file-like obj"""
+            return b"".join(up_to_iter(size))
+
+    return FileLikeObj()
